pyproject.toml
# https://python-poetry.org/docs/pyproject/
[tool.poetry]
name = "decenter-ai.streamlit.app"
version = "3.0.0"
description = "Official decenter ai infra app"
authors = ["Nasfame <laciferin@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://decenter-ai.streamlit.app"
repository = "https://github.com/Decenter-AI/decenter-ai.streamlit.app"
documentation = "https://github.com/Decenter-AI/decenter-ai.streamlit.app"
keywords = ["ai", "decenter", "ai-infra"]


packages = []

classifiers = [
  "Topic :: Software Development :: Build Tools",
  "Programming Language :: Python :: 3.x",

  "Development Status :: 4 - Beta",
  "Environment :: Web Environment",
  "Framework :: Streamlit",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: Your License Here",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Internet :: WWW/HTTP",
  "Topic :: Software Development :: Libraries :: Application Frameworks",
  "Topic :: System :: Distributed Computing",
  "Topic :: System :: Networking",
]


[tool.poetry.dependencies]
python = "^3.10"
streamlit = "1.26.0"
black = "^23.9.1"
clarifai = "^9.9.0"
clarifai-grpc = "^9.8.2"
colorama = "^0.4.6"
colored = "^2.2.3"
gtts = ">=2.3.2,<2.4.0"
icecream = "^2.1.3"
joblib = "^1.3.2"

openai = "^0.28.0"
pinecone-client = "^2.2.2"
pytest = ">=7.4.0,<7.5.0"
python-dotenv = "1.0.0"
pychroot = "0.9.18"
dataclasses-json = "^0.6.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
pytest-benchmark = "^4.0.0"
flake8-in-file-ignores = "^0.2.2"
pre-commit = "^3.4.0"

[tool.poetry.group.ml.dependencies]
numpy = "^1.25.2"
pandas = "^2.1.0"
requests = ">=2.31.0,<2.32.0"
scikit-learn = "^1.3.0"
jupyter = "^1.0.0"
matplotlib = "^3.8.0"
ipykernel = "^6.25.2"
beautifulsoup4 = "^4.12.2"
seaborn = "^0.13.0"
spacy = "^3.6.1"
statsmodels = "^0.14.0"
pillow = "9"
tensorflow-addons = "^0.21.0"
regex = "^2023.8.8"


[tool.poetry.group.tensorflow-addons.dependencies]
regex = "^2023.10.3"

[build-system]
# requires = ["python>=3.10", "setuptools", "wheel"]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]


[tool.poetry.extras]
dev = ['pytest']
ml = [
  "numpy",
  "pandas",
  "scikit-learn",
  "matplotlib",
  "seaborn",
  "tensorflow",
  "keras",
  "pytorch",
  "spacy",
]

# A list of all of the optional dependencies, some of which are included in the
# poetry install --extras
# poetry install --all-extras

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/Decenter-AI/decenter-ai.streamlit.app/issues"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
